pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "feeld-local"
version = "0.1.0"
description = "Local Feeld client — personal use only"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "flask>=3.0",
    "urwid>=2.6",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.scripts]
feeld = "feeld.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["feeld*", "web*", "tui*"]
